requirements.txt
absl-py==2.3.1
astunparse==1.6.3
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
comtypes==1.4.13
contourpy==1.3.2
cycler==0.12.1
Flask==3.0.3
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
google-auth==2.41.1
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
grpcio==1.75.1
gunicorn==22.0.0
h5py==3.15.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.1.0
numpy==1.26.4
oauthlib==3.3.1
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==12.0.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
Pygments==2.19.2
pyparsing==3.2.5
pypiwin32==223
python-dateutil==2.9.0.post0
pyttsx3==2.99
pywin32==311
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
scipy==1.15.3
six==1.17.0
SpeechRecognition==3.10.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-estimator==2.15.0
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.0.3
wrapt==1.14.2
